refactor : 변수명 통일화 (~Value)
